websocket/requirements.txt
flask
flask_sock
mysql.connector
